# docs/api/openapi-query.yaml
openapi: 3.1.0
info:
  title: siglake query
  description: |-
    DataFusion SQL over the Iceberg tables, index and index-template management, batch jobs, GDPR delete tasks, and a Jaeger-compatible trace read API.

    Ingest lives on a separate server (`openapi-ingest.yaml`) — this one only reads, apart from index management and delete tasks.
  license:
    name: Apache-2.0
  version: 0.1.0
servers:
- url: http://localhost:8089
  description: Default `--bind` address
paths:
...
  /api/v1/jobs/{id}:
    get:
      tags:
      - jobs
      summary: Poll a batch job's status.
      description: |-
        Batch jobs are created by `POST /api/v1/sql` with `priority: "batch"`.
        Any replica answers for any job: the store is the catalog database by
        default (`query.jobs.persistent`), so the id is not tied to the pod that
        took the submission. With that turned off each pod keeps its own in-memory
        jobs and a poll routed elsewhere answers `404`.
      operationId: status
      parameters:
      - name: id
        in: path
        description: Job id (UUIDv7).
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Job state. `cost` is populated once planning has run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobInfo'
        '400':
          description: Malformed job id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '404':
          description: No such job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '503':
          description: The job store is temporarily unavailable; retry the request.
          headers:
            retry-after:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
    delete:
      tags:
      - jobs
      summary: Cancel a batch job.
      description: |-
        The `202` means the cancellation is *persisted* and terminal: the job row
        reads `cancelled` from every replica, and a completion the executor writes
        afterwards is refused. Pending or running execution is then aborted,
        releasing its query-admission reservation and cancelling its storage scans.

        When the executing replica is this one, that abort is immediate. When it is
        another replica — the ordinary case, since the store is shared by default
        and the chart runs two query pods — the abort handle lives in that process,
        and it observes the persisted cancellation within `--jobs-cancel-poll-secs`
        (default 2 s).
        So `202` bounds when the work stops rather than promising it has already
        stopped. A job that already reached a terminal state answers `409`.
      operationId: cancel
      parameters:
      - name: id
        in: path
        description: Job id (UUIDv7).
        required: true
        schema:
          type: string
      responses:
        '202':
          description: Cancellation persisted. The row is `cancelled` for every replica and a later completion is refused; the executing replica aborts the work within `--jobs-cancel-poll-secs` (default 2 s) when it is not the replica that served this request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCancelResponse'
        '400':
          description: Malformed job id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '404':
          description: No such job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '409':
          description: The job already reached a terminal state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '503':
          description: The job store is temporarily unavailable; retry the request.
          headers:
            retry-after:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
  /api/v1/jobs/{id}/result:
    get:
      tags:
      - jobs
      summary: Fetch a finished batch job's rows.
      description: |-
        Only valid once the job has succeeded: a still-running job answers `404` and
        a job that ended in any non-success state answers `409` — poll
        `GET /api/v1/jobs/{id}` to tell those apart.
      operationId: result
      parameters:
      - name: id
        in: path
        description: Job id (UUIDv7).
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Result rows.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordsResponse'
        '400':
          description: Malformed job id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '404':
          description: No such job, or it is still pending or running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '409':
          description: The job ended in a non-success state (failed, cancelled or timed out).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '503':
          description: The job store is temporarily unavailable; retry the request.
          headers:
            retry-after:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
  /api/v1/sql:
    post:
      tags:
      - sql
      summary: |-
        `POST /api/v1/sql` entry point. Transparently coordinates a distributed
        query when worker peers are configured and this is an ordinary interactive
        request; otherwise (no peers, a worker sub-request with `shard`, a batch
        job, or a dry-run) it runs the single-pod path, which is also always
        reachable at `POST /api/v1/sql/local`.
      operationId: handle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SqlRequest'
        required: true
      responses:
        '200':
          description: 'Success. Normally result rows (`RecordsResponse`); a `dry_run` request returns only the `CostReport` instead. When the request asks for `format: "ndjson"` the body is `application/x-ndjson` — one JSON object per line, streamed, with a trailing `{"_meta": ...}` line if the result was cut short or execution failed after the 200 response started. An error trailer carries `code` and `error`; a pool-refusal trailer also carries `retry_after_secs`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SqlSuccess'
            application/x-ndjson:
              schema:
                type: string
        '202':
          description: 'Admitted and queued on the dedicated runtime when the request set `priority: "batch"`. Poll the returned URLs. Batch submissions share the query admission budget and return 429 instead when it remains full through the admission wait.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSubmitResponse'
        '400':
          description: Empty or unparseable SQL, a statement that is not a read (DDL, DML and session statements — `CREATE`, `COPY … TO`, `INSERT`, `SET` — are refused before they can take effect), or a rejected cost estimate. `cost` carries the estimate that was refused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '413':
          description: 'Two distinct cases. Either the row cap was hit while rendering — the body is then a `RecordsResponse` with `truncated: true` and the rows that did fit — or the mid-flight rows-scanned breaker tripped, in which case the body is an error carrying `cost` and `stats`. Distinguish by whether `rows` is present.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '422':
          description: 'Unprocessable request body: a key inside `limits` is not a known per-request limit. The commonest case is `max_rows`, which is the RESPONSE envelope''s name for the applied cap; the request field is `max_rows_returned`. Refused by the JSON extractor before planning, execution or batch enqueue, so no work is done and no job is created — a dropped cap is never applied silently. The body is the extractor''s plain-text message naming the unknown field, not an `ApiErrorBody`. Unknown keys at the TOP level of the request are still ignored.'
        '429':
          description: Admission control rejected the interactive query or batch submission; the server is at its in-flight ceiling.
          headers:
            retry-after:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '503':
          description: 'Retryable refusal, from this server or from a worker whose answer is forwarded: the query memory pool (or the spill directory''s byte cap) refused an allocation, or a fanned-out shard could not resolve the generation — table incarnation, snapshot or schema — the coordinator pinned it to (`reason: "shard_pin_unresolved"`), which is refused rather than answered from a mix of generations. Nothing is wrong with the query: wait `Retry-After` seconds and retry.'
          headers:
            retry-after:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '504':
          description: The query exceeded its wall-clock budget.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
...
components:
  schemas:
    ApiErrorBody:
      type: object
      description: |-
        Error body used by every endpoint on this server.

        Spec-only. Some errors additionally flatten diagnostic keys into the top
        level — `cost` on a rejected or timed-out query, and `cost` plus `stats` on
        the mid-flight rows breaker — so this schema deliberately allows additional
        properties rather than pinning an exact shape.
      required:
      - error
      - code
      properties:
        code:
          type: integer
          format: int32
          description: Echo of the HTTP status code.
          minimum: 0
        cost:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/CostReport'
            description: Present when the failure was cost- or time-related.
        error:
          type: string
          description: Human-readable failure description.
...
    BatchSubmitResponse:
      type: object
      description: |-
        `202 Accepted` body returned by `POST /api/v1/sql` when the request carries
        `priority: "batch"`: after admission, the query is queued rather than
        executed inline and the caller polls the returned URLs. A full shared query
        admission budget returns `429` instead of creating a job.
      required:
      - job_id
      - status_url
      - result_url
      - priority
      properties:
        job_id:
          type: string
          description: Job identifier (UUIDv7, so ids sort by submission time).
        priority:
          type: string
          description: Always `batch`.
          example: batch
        result_url:
          type: string
          description: |-
            Path to fetch the result once the job succeeds —
            `GET /api/v1/jobs/{id}/result`.
        status_url:
          type: string
          description: Path to poll for status — `GET /api/v1/jobs/{id}`.
...
    JobCancelResponse:
      type: object
      description: '`202 Accepted` body of `DELETE /api/v1/jobs/{id}`.'
      required:
      - cancelled
      properties:
        cancelled:
          type: string
          description: Id of the job that was cancelled.
    JobInfo:
      type: object
      description: |-
        Public snapshot of job state (omits the result body — fetch that
        separately via `/result`).
      required:
      - job_id
      - status
      - priority
      - submitted_at
      - query
      properties:
        cost:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/CostReport'
        ended_at:
          type:
          - string
          - 'null'
          format: date-time
        error:
          type:
          - string
          - 'null'
        job_id:
          type: string
        priority:
          $ref: '#/components/schemas/Priority'
        query:
          type: string
        started_at:
          type:
          - string
          - 'null'
          format: date-time
        status:
          $ref: '#/components/schemas/JobStatus'
        submitted_at:
          type: string
          format: date-time
    JobStatus:
      type: string
      enum:
      - pending
      - running
      - succeeded
      - failed
      - cancelled
      - timeout
...
    Priority:
      type: string
      description: Query priority — picks which tier's defaults + ceilings apply.
      enum:
      - interactive
      - batch
...
    RecordsResponse:
      type: object
      description: JSON envelope for `QueryFormat::Records`.
      required:
      - columns
      - row_count
      - rows
      properties:
        approximation:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Approximation'
            description: |-
              Present ONLY when the answer is approximate, and then always — with the
              bound and the residual, never a bare flag.

              Absence means exact. That asymmetry is deliberate: a caller that ignores
              this field entirely still gets exact answers treated as exact, and the
              only way to be misled is to receive this field and discard it.
        columns:
          type: array
          items:
            type: string
        cost:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/CostReport'
            description: |-
              Cost report attached to every successful response so callers
              can track effort without a separate `/explain` round trip.
        max_rows:
          type:
          - integer
          - 'null'
          description: |-
            Effective row cap that produced the truncation. Only present
            when `truncated == true`.
          minimum: 0
        row_count:
          type: integer
          minimum: 0
        rows:
          type: array
          items:
            type: object
          description: One JSON object per row, keyed by `columns`.
        stats:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ScanStats'
            description: |-
              Per-query leaf scan accounting (rows/bytes read, spill). Set by the
              execution path; `Some(ScanStats::default())` on fast-path responses.
        truncated:
          type: boolean
          description: True when the row cap was hit and the result is incomplete.
...
    SqlSuccess:
      oneOf:
      - $ref: '#/components/schemas/RecordsResponse'
        description: Result rows. The normal case.
      - $ref: '#/components/schemas/CostReport'
        description: 'Cost estimate only, returned when the request set `dry_run: true`.'
      description: |-
        The two success bodies `POST /api/v1/sql` can return at `200`: a dry run
        answers with only a cost report, while a real query answers with result rows.
  securitySchemes:
    query_token:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <token>`. With OIDC configured the token is a JWT and its claims select the tenant; otherwise it is matched against the cluster''s static token list. A server started with neither accepts unauthenticated requests.'
security:
- query_token: []
...